+++ components/ui/keyboard/Key.tsx
@@ -1,260 +1,115 @@
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faUp } from "@fortawesome/pro-regular-svg-icons";
-import { faDeleteLeft } from "@fortawesome/pro-solid-svg-icons";
-import { faCircle } from "@fortawesome/pro-solid-svg-icons";
-import { faArrowTurnDownLeft } from "@fortawesome/pro-solid-svg-icons";
+import { CSSProperties, useEffect, useState } from "react";
 
-import styles from "../../../styles/CustomKeyboard.module.scss";
+import styles from "../../../styles/keyboard/Key.module.scss";
 
-import { KEY_ACTION } from "./CustomKeyboard";
 import { Variables } from "../../../../variables";
-import { CSSProperties, useState } from "react";
+
+import { IInputAction } from "../../../interfaces";
+import { IKeyOptions, KEY_ACTION } from "../../../lib/keyboardTypes";
+
+import { useKeyboardContext } from "../../../contexts/keyboardContext";
+
+import ActionKey from "./ActionKey";
 
 interface IKeyProps {
-	text?: {
-		defaultValue: string
-		capslockValue?: string
-		specCharsValue?: string
+	index: number;
+	config: IKeyOptions
+	actionsOverride?: {
+		[keyIndex: string]: IInputAction[]
 	}
-	style?: CSSProperties
-	action?: KEY_ACTION | CallableFunction
-	onChangeText: CallableFunction
-	onDeleteText?: CallableFunction
-	onShiftToggle?: CallableFunction
-	capslock?: boolean
-	onDisplaySpecChars?: CallableFunction
-	specChars?: boolean
+	styleOverride?: {
+		index: number | "all"
+		style: CSSProperties
+		valueOverride?: string
+	}[]
 }
 
 export default function Key(props: IKeyProps): JSX.Element {
-	const {
-		text,
-		style,
-		action,
-		onChangeText,
-		onDeleteText,
-		onShiftToggle,
-		capslock,
-		onDisplaySpecChars,
-		specChars,
-	} = props;
+	const { index, config, actionsOverride, styleOverride, } = props;
+
+	const [classNames, setClassNames] = useState<string[]>([styles.default]);
+	const [text, setText] = useState<string>("");
+
+	const [customActions, setCustomActions] = useState<IInputAction[]>([]);
+	const [customStyles, setCustomStyle] = useState<CSSProperties>();
 
 	const [pressed, setPressed] = useState<boolean>(false);
 
-	const changetexthandler = () => {
-		if (capslock && text?.capslockValue !== undefined) {
-			onChangeText(text.capslockValue);
-		} else if (specChars && text?.specCharsValue !== undefined) {
-			onChangeText(text.specCharsValue);
+	const { capslock, specChars, onChange, triggerActionsOverride, } = useKeyboardContext();
+
+	useEffect(() => {
+		if (pressed) {
+			setClassNames(latest => [...latest, styles.pressed]);
 		} else {
-			onChangeText(text?.defaultValue);
+			setClassNames(latest => {
+				const returned = latest.filter(name => name !== styles.pressed);
+				return [...returned];
+			});
 		}
-	};
+	}, [pressed]);
 
-	function changetexthandlerDev() {
-		if (Variables.PREVIEW) {
-			changetexthandler();
+	useEffect(() => {
+		if (capslock && config.text?.capslockValue) {
+			setText(config.text.capslockValue);
+		} else if (specChars && config.text?.specCharsValue) {
+			setText(config.text.specCharsValue);
+		} else {
+			setText(config.text?.defaultValue ?? (config.action === KEY_ACTION.SPACEBAR ? "Space" : ""));
 		}
-	}
 
-	const displaySpecCharsHandler = () => {
-		if (onDisplaySpecChars) {
-			onDisplaySpecChars();
-		}
-	};
-
-	function displaySpecCharsHandlerDev() {
-		if (Variables.PREVIEW) {
-			displaySpecCharsHandler();
-		}
-	}
-
-	if (action) {
-		if (action === KEY_ACTION.SHIFT) {
-			const shiftClickHandler = () => {
-				if (onShiftToggle) {
-					onShiftToggle();
+		if (styleOverride) {
+			styleOverride.map(style => {
+				if (style.index === "all" || style.index === index) {
+					setText(latest => style.valueOverride ?? latest);
+					setCustomStyle(latest => {
+						return {
+							...latest,
+							...style.style,
+						};
+					});
 				}
-			};
-
-			const devClick = () => {
-				if (Variables.PREVIEW) {
-					shiftClickHandler();
-				}
-			};
-
-			return (
-				<div
-					className={`${styles.key_action} ${pressed ? styles.pressed : ""}`}
-					onClick={devClick}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						shiftClickHandler();
-						setPressed(false);
-					}}
-				>
-					{text ? text.defaultValue : <FontAwesomeIcon icon={faUp} style={{ color: "#999999", fontSize: "0.03rem", }} />}
-					<FontAwesomeIcon icon={faCircle} style={{ color: capslock ? "#00dd00" : "#999999", fontSize: "0.01rem", marginTop: "0.004rem", }} />
-				</div>
-			);
+			});
 		}
+	}, [capslock, specChars, styleOverride]);
 
-		if (action === KEY_ACTION.SPECIALCHARS) {
-			return (
-				<div
-					className={`${styles.key_action} ${pressed ? styles.pressed : ""}`}
-					onClick={displaySpecCharsHandlerDev}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						displaySpecCharsHandler();
-						setPressed(false);
-					}}
-				>
-					<p>{text ? text.defaultValue : "#+="}</p>
-					<FontAwesomeIcon icon={faCircle} style={{ color: specChars ? "#00dd00" : "#999999", fontSize: "0.01rem", marginTop: "0.004rem", }} />
-				</div>
-			);
+	useEffect(() => {
+		if (actionsOverride && actionsOverride[index]) {
+			setCustomActions(actionsOverride[index]);
 		}
+	}, [actionsOverride]);
 
-		if (action === KEY_ACTION.SPACEBAR) {
-			return (
-				<div
-					className={`${styles.key_spacebar} ${pressed ? styles.pressed : ""}`}
-					onClick={() => Variables.PREVIEW ? onChangeText(" ") : ""}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						onChangeText(" ");
-						setPressed(false);
-					}}
-				>
-					<p>{text ? text.defaultValue : "Space"}</p>
-				</div>
-			);
-		}
-
-		if (action === KEY_ACTION.ENTER) {
-			return (
-				<div
-					className={`${styles.key_action} ${pressed ? styles.pressed : ""}`}
-					onClick={() => Variables.PREVIEW ? onChangeText("\n") : ""}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						onChangeText("\n");
-						setPressed(false);
-					}}
-					style={style}
-				>
-					<p
-						style={{
-							color: style?.color ? style.color : "",
-							fontSize: style?.fontSize ? style.fontSize : "",
-						}}
-					>
-						{text ? text.defaultValue : <FontAwesomeIcon icon={faArrowTurnDownLeft} style={{ color: "#999999", fontSize: "0.02rem", }} />}
-					</p>
-				</div>
-			);
-		}
-
-		if (action === KEY_ACTION.BACKSPACE) {
-			return (
-				<div
-					className={`${styles.key_action} ${pressed ? styles.pressed : ""}`}
-					style={style}
-					onClick={() => {
-						if (Variables.PREVIEW) {
-							if (onDeleteText) {
-								onDeleteText();
-							}
-						}
-					}}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						setPressed(false);
-						if (onDeleteText) {
-							onDeleteText();
-						}
-					}}
-				>
-					{text ? text.defaultValue : <FontAwesomeIcon icon={faDeleteLeft} style={{ color: "#999999", fontSize: "0.025rem", }} />}
-				</div>
-			);
-		}
-
-		if (typeof action === "function") {
-			let displayedText = text?.defaultValue;
-
-			if (capslock && text?.capslockValue) {
-				displayedText = text.capslockValue;
-			}
-			if (specChars && text?.specCharsValue) {
-				displayedText = text.specCharsValue;
-			}
-
-			return (
-				<div
-					className={`${styles.key_action} ${pressed ? styles.pressed : ""}`}
-					onClick={() => Variables.PREVIEW ? action() : ""}
-					onMouseDown={() => setPressed(true)}
-					onMouseUp={() => setPressed(false)}
-					onTouchStart={() => setPressed(true)}
-					onTouchEnd={() => {
-						action();
-						setPressed(false);
-					}}
-					style={{
-						backgroundColor: style?.backgroundColor ? style.backgroundColor : "",
-						justifyContent: style?.textAlign ? style.textAlign : "",
-					}}
-				>
-					<p
-						style={{
-							color: style?.color ? style.color : "",
-							fontSize: style?.fontSize ? style.fontSize : "",
-						}}
-					>
-						{displayedText ?? "Action"}
-					</p>
-				</div>
-			);
+	function triggerActions() {
+		if (customActions.length > 0 && triggerActionsOverride) {
+			triggerActionsOverride(customActions);
 		}
 	}
 
-	let displayedText = text?.defaultValue;
-
-	if (capslock && text?.capslockValue) {
-		displayedText = text.capslockValue;
+	if (config.action) {
+		return <ActionKey config={{ index, config, customStyles, customText: text, }} onTriggerActionsOverride={triggerActions} />;
 	}
-	if (specChars && text?.specCharsValue) {
-		displayedText = text.specCharsValue;
+
+	function clickHandler() {
+		setPressed(false);
+		onChange(text);
+		triggerActions();
+	}
+
+	function devClick() {
+		if (Variables.PREVIEW) {
+			clickHandler();
+		}
 	}
 
 	return (
 		<div
-			className={`${styles.key_default} ${pressed ? styles.pressed : ""}`}
-			style={style}
-			onClick={changetexthandlerDev}
-			onMouseDown={() => setPressed(true)}
-			onMouseUp={() => setPressed(false)}
+			className={classNames.join(" ")}
+			style={customStyles}
 			onTouchStart={() => setPressed(true)}
-			onTouchEnd={() => {
-				changetexthandler();
-				setPressed(false);
-			}}
+			onTouchEnd={clickHandler}
+			onMouseDown={() => setPressed(true)}
+			onMouseUp={devClick}
 		>
-			<p>
-				{displayedText}
-			</p>
+			<p>{text}</p>
 		</div>
 	);
 }
